refactor(Host,Race): Host,Race 클래스 리팩토링
기존의 Race 클래스는 게임의 전체적인 흐름을 주관하여 시도횟수도 이곳에서 입력받았다. 이 부분에서 시도횟수 필드인 attemptNumber가 불변하지 않다는 것이 옳지 않다고 생각하여 Host 클래스를 생성 후 리팩토링을 진행하였다.
