+++ src/main/java/racingcar/Application.java
@@ -6,7 +6,7 @@ public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
         // TODO 구현 진행
-        Race race = new Race(scanner);
-        race.start();
+        Host host = new Host(scanner);
+        host.start();
     }
 }
